+++ app/src/main/java/com/interiordesign3d/ui/screen/splash/state/SplashState.kt
@@ -1,0 +1,7 @@
+package com.interiordesign3d.ui.screen.splash.state
+
+import androidx.compose.runtime.Stable
+import com.interiordesign3d.common.base.BaseScreenState
+
+@Stable
+open class SplashState : BaseScreenState()